Move product extents tool under `datacube product` cli
